+++ libft/ft_min.c
@@ -6,13 +6,13 @@
 /*   By: dwinky <dwinky@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/04 15:03:57 by dwinky            #+#    #+#             */
-/*   Updated: 2021/01/04 19:56:59 by dwinky           ###   ########.fr       */
+/*   Updated: 2021/01/06 21:28:52 by dwinky           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_min(long long x, long long y)
+long long	ft_min(long long x, long long y)
 {
 	return (x < y ? x : y);
 }
